requirements.txt
fastapi
uvicorn
farm-haystack[elasticsearch,crawler]
python-multipart
sqlalchemy
setuptools
